requirements.txt
apiclient==1.0.4
asgiref==3.2.5
auth==0.5.3
blinker==1.4
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-stubs==1.1.0
dnspython==1.16.0
ecdsa==0.15
eventlet==0.25.2
falcon==2.0.0
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.13.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
greenlet==0.4.15
gunicorn==20.0.4
httplib2==0.17.1
idna==2.9
mongoengine==0.19.1
monotonic==1.5
oauth2client==4.1.3
oauthlib==3.1.0
Pillow==7.0.0
protobuf==3.11.3
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.10.1
python-jose==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
Werkzeug==1.0.1

